+++ src/components/editor/markdown-renderer/markdown-it-plugins/highlighted-code.ts
@@ -1,0 +1,23 @@
+import MarkdownIt from 'markdown-it/lib'
+
+const highlightRegex = /^(\w*)(=?)$/
+
+export const highlightedCode: MarkdownIt.PluginSimple = (md: MarkdownIt) => {
+  md.core.ruler.push('highlighted-code', (state) => {
+    state.tokens.forEach(token => {
+      if (token.type === 'fence') {
+        const highlightInfos = highlightRegex.exec(token.info)
+        if (!highlightInfos) {
+          return
+        }
+        if (highlightInfos[1]) {
+          token.attrJoin('data-highlight-language', highlightInfos[1])
+        }
+        if (highlightInfos[2]) {
+          token.attrJoin('data-show-gutter', '')
+        }
+      }
+    })
+    return true
+  })
+}