+++ CHANGELOG.md
@@ -16,6 +16,11 @@
     - If a legacy embedding code is detected it will show the link to the presentation instead of the embedded presentation
 - Speakerdeck embedding
     - If a legacy embedding code is detected it will show the link to the presentation instead of the embedded presentation
+- We are now using `highlight.js` instead of `highlight.js` + `prism.js` for code highlighting. Check out the [highlight.js demo page](https://highlightjs.org/static/demo/) to see which languages are supported.
+ The highlighting for following languages isn't supported by `highlight.js`:
+    - tiddlywiki
+    - mediawiki
+    - jsx
 
 ### Added